requirements.txt
psutil
pymediainfo
python-dateutil
pyudev
pyyaml
requests
watchdog
